[PORT FROM MAIN] WIDI Connectivity: Sort WIFI displays in alphabetic order
BZ: 84083

In WIFI displays list, the devices are sorted by RSSI level and thus move
according to environment. The refresh occurs so often that when clicking on one
device while the list is being updated, an undesired device can be choose.

To fix this, this patch proposes to displays devices in alphabetic order.

Change-Id: I96f15445b3326c7624b108cfb52ed6de06815cfb
Orig-Change-Id: I922da6ed94f188e0b11a081a822160d44c555518
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/123541
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>