EM/THERMAL: Enable dethrottling in warning state for CPU zone
BZ: 86634

This patch enables the CPU dethrottling when the CPU zone
moves to normal state. This was disabled as dethrottling was
done by GPU skin but this was causing CPU freq to stuck in
throttled state. Change made to default thermal config xml located
in core/res/res/xml folder.

Change-Id: I114804591885c3992d63a7d19f02a88fa236b115
Signed-off-by: Saikat <saikat.roychowdhury@intel.com>
Reviewed-on: http://android.intel.com:8080/93211
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: B M, Shravan K <shravan.k.b.m@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>