ActivityManagerService : Dump dropbox logfile content to Slog when dropbox full
BZ: 89799

When dropbox directory is full, the dropbox manager only produces zero-length
log files to not take supplementary space on data partition.

In order to retrieve debug data when an ANR or a System_Server Watchdog happens
in full dropbox conditions, this patch allows to flush the content of the stack
traces file (if not null) to the System log before it is passed to the
DropBoxManagerService.

Change-Id: I30944b6c3690331412e0e40a6b976ef0d05e4956
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/94135
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>