[PORT FROM MAIN] WiFi: Do not scan in background when P2P connected.
BZ: 84417

When no WiFi network is remembered, wpa_supplicant do not
automatically schedule periodic scan.
In that case, Android framework take the action to trig a scan
with the same interval of time than "suplicant scan interval".
(Intel set it to 40s)
When P2P is connected, this Android framework scan is not triggered.
And supplicant scan interval is changed.
Intel set it to 3600s in order to minimize scan impact on P2P
connection, mainly for Miracast purpose.

Change-Id: Ifa29f3e134ad33d01c8a9dce7d0a4f33cacbea62
Orig-Change-Id: I54625640d7217458d1da5469848adf6baf016565
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: http://android.intel.com:8080/123406
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>