[PORT FROM MAIN] NetworkController: display the carrier name even when there is no data connection
BZ: 81455

The carrier name is replaced by 'NO INTERNET CONNECTION' when no
data connection available. It was requested by CCB to change this
behavior to have the carrier name displayed in all cases. This is
requested for some operator.

This patch changes the default behavior of the Android/UI to have
the carrier name displayed even when there is no data connection.
If the carrier name is empty the 'NO INTERNET CONNECTION' message
is displayed.

Orig-Change-Id: I168117ad1b04e842af787f72828ac01ba465554c
Category: aosp improvement
Domain: AOSP-Framework-SystemUI
Origin: internal
Upstream-Candidate: yes
Change-Id: I5dc7e13d2a3ebfc7b930001e38314a3c19100e33
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>