[BT] Improve airplane mode handling
BZ: 84907

With this patch, AIRPLANE_MODE_CHANGED intents will be
handled by BT only when it is in a stable state (ON or OFF).
If multiple intents regarding a change in the airplane mode
state are received while BT is turning on or turning off,
only the last intent will be handled.
This patch is necessary to avoid BT being overwhelmed by too
many state changes in a short period of time (like 50 airplane
mode changes in a few seconds) as this would often lead to a
crash.

Change-Id: I9f1170cce49a2f43c429974e6b439e047ad571b3
Signed-off-by: Thomas Lehaux <thomasx.lehaux@intel.com>
Signed-off-by: Cedric Bondier <cedricx.bondier@intel.com>
Reviewed-on: http://android.intel.com:8080/96744
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Boniface, XavierX <xavierx.boniface@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>