[PORT FROM R42S] keyguard: Update sim state on successful SIM unlock
BZ: 88437

After entering PUK and new PIN code, a numeric
keypad briefly appears without any other keys.
Its appearance/disappearance is sudden enough for
the user to be unable to do anything with it.

Once the request for unlock using PUK has been
completed successfully, showNextSecurityScreenOrFinish
is called to finish showing the keyguard. Since the
security mode is still in SimPuk mode instead of
dismissing the keyguard, new security screen again.
This is because sim state is not updated after the
SIM PUK is handled successfully.

After the SIM PUK unlock is handled successfully,
update the sim state before requesting for keyguard
dismiss.

Change-Id: I5e700744ed866d4c7770f5d554f0d46c9f4c1f2f
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/102048
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>