EM/THERMAL: Fix crash when zone added without Cooling device attached
BZ: 105792

This patch fixes the crash that was being observed when a zone is
added to the thermal_throttle_config.xml file without having any
cooling device attached to the zone. This was earlier forcing us
to have a cooling device for every zone even if no action was
being taken on the cooling device. Not an expected behaviour.

Change-Id: I480c29195f045e2f8cc90efda117fed28b8c2f34
Signed-off-by: Ajay <ajay.thomas.david.rajamanickam@intel.com>
Reviewed-on: http://android.intel.com:8080/106291
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Roychowdhury, Saikat <saikat.roychowdhury@intel.com>
Reviewed-by: B M, Shravan K <shravan.k.b.m@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: R, Durgadoss <durgadoss.r@intel.com>
Reviewed-by: Mariappan, Bhagat SinghX <bhagat.singhx.mariappan@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>