LEX Upstream: Spell Checker Memory Leak
BZ: 106704

LEX Upstream:
Created to apply the fix found in LEX product in PPD to main branch.
There was memeory leak reported in LEX product and there is some problem in Spell check session.
When spell checker served-view exited, its reference is always held by spellcheck session listener.
So it should be initialized as NULL when close function is called.

Change-Id: I13356d85b804759c522d7afcee9df3d5fb7ae6c8
Reviewed-on: http://android.intel.com:8080/106862
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Papadkar, Umesh N <umesh.n.papadkar@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Shentu, JiazhenX <jiazhenx.shentu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>