[PORT FROM MAIN] ASF 1.0 initial Android system services changes
BZ: 108725

ASF relies on custom modifications to certain Android system services
such as the Activity Manager and the Package Manager, as well as
changes to the core AndroidManifest.xml to add the ASF permissions
to the system.

Coding style changes, replace tab with space for ASF code changes in
Activity Manager and the Package Manager.

Coding style changes, remove trailing space.

Orig-Change-Id: Ic461a0c681b5cd6b9531ee833861873e947dc55e
Signed-off-by: Wai Ming Lee <waix.m.lee@intel.com>

Conflicts:

	services/java/com/android/server/am/ActivityManagerService.java

Change-Id: I14c1dba3359dded8d0c4e8c5620c864857d8aabb
Reviewed-on: http://android.intel.com:8080/123643
Reviewed-by: Shuster, Norvin <norvin.shuster@intel.com>
Reviewed-by: Gopalaiah, Chandra Prakash <chandra.prakash.gopalaiah@intel.com>
Reviewed-by: Allyn, Mark A <mark.a.allyn@intel.com>
Tested-by: Allyn, Mark A <mark.a.allyn@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>