HTML5: When enter to full screen mode, get video height and width
BZ: 103140

Before View::onMeasure is called, we need video height and width to
establish the real video display size or else it will be set to
default video surface size which equal to the screen size. So the
video frame seems to be stretched caused by the wrong video height.

Change-Id: I0740bf58d682ebbc504cd2ed4a8c7e96ca6ba8a1
Signed-off-by: Xiao,Feng <fengx.xiao@intel.com>
Reviewed-on: http://android.intel.com:8080/108456
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ji, WeiweiX <weiweix.ji@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Wang, Yi A <yi.a.wang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>