[PORT FROM R42B-STABLE] Reboot in COS if charger plugged
BZ: 109667

The board shall reobot in COS if a charger is plugged and
user requested a shutdown.
This was previously done in kernel, this patch is providing this in userland (Java).

This behavior is managed by the presence of a specific Android property

Change-Id: Ie22a4727df82717fcf97cc38aa8e162b36e11faf
Orig-Change-Id: I16baec297f742d884524e843533e7797fd6073bf
Category: feature differentiation
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: no
Signed-off-by: Frederic Despalles <frederic.fespalles@intel.com>