Random Browser Force close from SpellCheckSession.
BZ:110617

The following issue was reported from LEX product team. This cannot be reproduced easily at normal situation
but it is a timing related issue and if system gets slow or in stress test, this force close could happen.
In SpellCheckerSession.close()in SpellCheckerSession.java, mSpellCheckerSessionListener needs to be initialied
to NULL after mSpellCheckerSessionListenerImpl.close is executed. Otherwise, there is the chance that
handleOnGetSuggestionsMultiple is called after mSpellCheckerSessionListener is assigned to NULLL but
mSpellCheckerSessionListenerImpl.close() is not finished as below.
private void handleOnGetSuggestionsMultiple(SuggestionsInfo[] suggestionInfos) {
        mSpellCheckerSessionListener.onGetSuggestions(suggestionInfos);
}

Change-Id: I0ef5267ac45889258f9e2e8c6d7a1fd754d3c3ca
Reviewed-on: http://android.intel.com:8080/109531
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Byun, Seokjae <seokjae.byun@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>