Add debug info to provide wdt timing and term info
BZ: 110809

Added a class member representing the last time we started to monitor
an object so that the last monitored object timing will be know even
if not terminated.

Also check a last time that the monitor is terminated or not because
we are facing many cases where it seems the monitor terminates while
the stack of interest are dumped and then showing that the
ServerThread is currently doing nothing.

Change-Id: Ia259f1ab270547a9ad9bd45c29fb27a5ab2b658a
Signed-off-by: Jean-Christophe PINCE <jean-christophe.pince@intel.com>
Reviewed-on: http://android.intel.com:8080/110025
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Reviewed-by: Zhang, Guobin <guobin.zhang@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>