policy: Fix for SIM pin dialog shown on sim not ready
BZ: 107119

[DO NOT UPSTREAM]

When the airplane mode is activated, sim pin dialog
is shown even when the sim state is transitioned from
pin to not ready. When the sim is not accessible in
radio off, sim state changed intent can be received
with sim not ready. Since sim not ready is not handled,
sim pin dialog is shown even when the SIM is not
accessible in radio off state.

Handle the ACTION_SIM_STATE_CHANGED intent received
with INTENT_VALUE_ICC_NOT_READY.

Change-Id: I8cedf8a2ec45ee9e821c0bdda722f3803f722fb4
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/110413
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>