HTML5 Streaming: Hide NavigationBar when controlBar disappered.
BZ: 115879

When playback html5 video streaming in full screen mode,we can see NavigationBar
all the time which causes more power consumer. Need to hide NavigationBar
for saving power.
This patch also includes the fix that not show media controller when surface changed
if it is not in showing status.

Signed-off-by: Feng Xiao <fengx.xiao@intel.com>

Orig-Change-Id: Ibaa2fb2fc57f951bd52a39ac0910d529639848fa
Change-Id: I6b1d8656b3a3d56feb3285cd523999559201c2ff
Reviewed-on: http://android.intel.com:8080/113306
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Xiao, FengX <fengx.xiao@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Guo, Nana N <nana.n.guo@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>