Enable Silent Switch Control into framework phone policy, only for SALTBAYE
BZ: 84164

The Application need to listen to the key event SW_MUTE from keypad driver( already provided by driver).
As it done for volume key up and down. Once event is detected mute/unmute action need to be taken.
This action should be on the same line as the user selecting the silent mode on the phone.

Change-Id: I853642524f3e6d2c4868c889cb00aad6577c9954
Signed-off-by: jshe32X<jianchunx.shen@intel.com>
Reviewed-on: http://android.intel.com:8080/113970
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Krishna, Samaga <samaga.krishna@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>