[PORT FROM MAIN]Enable Silent Switch Control into framework phone policy, only for SALTBAYE
BZ: 84164

The Application need to listen to the key event SW_MUTE from keypad driver( already provided by driver).
As it done for volume key up and down. Once event is detected mute/unmute action need to be taken.
This action should be on the same line as the user selecting the silent mode on the phone.

Change-Id: Ib07a0d43a0e068438327c98db932934368d1639e
Orig-Change-Id: I853642524f3e6d2c4868c889cb00aad6577c9954
Signed-off-by: jshe32X<jianchunx.shen@intel.com>
Reviewed-on: http://android.intel.com:8080/124845
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>