[PORT FROM MAIN] ASF 1.0: Fix uninstall/stop PackageInfo race
BZ: 112930

Under certain conditions, a package uninstall of a running application
will lead to a problem where the application stop event generator cannot
obtain the associated PackageInfo data, due to the package having been
removed from the system.

These changes to the frameworks/base project include the following:

1. All hook code is now guarded by a try/catch, so exceptions thrown
from the hook code (due to bugs in ASF) will not cause the package
manager or activity manager threads to terminate.

2. We now pass the pid of the stopped process to our sendAppStopEvent()
method, allowing it to associate the stop event with a recent uninstall
event.

NOTE: This patch must be installed in conjunction with the related patch
to the /vendor/intel/asf project.

Change-Id: Idf43f3069f84a2954e4f8b8871c98ba29ea16aa2
Orig-Change-Id: Ib1ef55524d083f269c30ae312fa50aca76dd34a4
Signed-off-by: Simmons, DavidX <davidx.l.simmons@intel.com>
Reviewed-on: http://android.intel.com:8080/123644
Reviewed-by: Shuster, Norvin <norvin.shuster@intel.com>
Reviewed-by: Allyn, Mark A <mark.a.allyn@intel.com>
Tested-by: Allyn, Mark A <mark.a.allyn@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>