[PORT FROM MAIN][Movie studio]: fix the messy thumbnail when imported one special clip(2)
BZ: 113431

It's android defect.
root cause:
When MP4 container doesn't have stss table it assumes all samples are
sync samples. So it's possible to seek to a p frame and see messy thumbnail.
In the transition case, from the design logic, we can't avoid getting a messy
thumbnail during the transition clip, so we use the black thumbnail to replace
the messy one.
In CTP/Merrifield platform, this patch work with http://android.intel.com:8080/#/c/113762/
In Baytrail platform, this patch work with http://android.intel.com:8080/#/c/113972/

Change-Id: I66671b79c893cabca2ba1447ec105bfbc6474c54
Orig-Change-Id: I1acbf46082776cc59f2863c24da849d9b95bb969
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-on: http://android.intel.com:8080/121770
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>