[REVERT ME] ShutdownThread: Increase radio switch off timer value
BZ: 119928

[DO NOT UPSTREAM]

Currently, there is a 12second timer to switch off the
radios(modem, NFC, BT). In corner cases, this 12seconds
timer is too less to find out the reason behind the radio
switch off delay. Even though the target is to switch off
the radio in 8seconds, it is better to increase the timer
value to get more information on radio not switching off
in 8seconds.
Also, in case of flashless modem, modem switch off takes
more than 12seconds.

Since there is a 30seconds timer in MMGR, increase the
radios switch off timer to 32seconds.

This patch should be reverted once the modem shutdown
happens properly within 8seconds(even in corner cases).

Change-Id: I581cf3dfa742e3d847661ed4879af8b0efd36371
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/116075
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>