[PORT FROM MAIN]wifi/tethering: Increase tether interface response timeout
BZ: 119401

WifiStateMachine is responsible to poll tether interface when DUT is used as
hotspot and if no response is coming back within 5 seconds, the state machine
closes the tether interface.

However, when many clients are connected as STA and are downloading a lot of
data (FTP for example) the interface can suffer from latency and doesn't respond
to the WifiStateMachine before timeout.

This patch increase the timeout from 5 to 10 seconds in order to test if it can
solve the issue.

Note: as a final solution, a smart timeout could be agile in order to fix the
delay depending of the number of connected STA.

Change-Id: I77c99d775da1279cd0359f2844b367b2ac534889
Orig-Change-Id: Ia47076bfa5ec67b60068833043708aa493d9d697
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no,proprietary
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/123900
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>