WiFi Hotspot: Increase tether interface response timeout
BZ: 119401

WifiStateMachine is responsible to poll tether interface when DUT is used as
hotspot and if no response is coming back within 5 seconds, the state machine
closes the tether interface.

However, when many clients are connected as STA and are downloading a lot of
data (FTP for example) the interface can suffer from latency and doesn't respond
to the WifiStateMachine before timeout.

This patch increase the timeout from 5 to 10 seconds in order to test if it can
solve the issue.

Note: as a final solution, a smart timeout could be agile in order to fix the
delay depending of the number of connected STA.

Change-Id: I42e00cbda0feb45587f34dc73adcd2fe8f58d627
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/116394
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>