Reset decorView for panel when device rotation
BZ: 97553

Rootcause:When device rotation, The ExpandMenuView and ViewRootImpl
will be re-created.But the decorView in PanelFeatureState don't update.
So the different ViewRootImpl will use the same decorView.
For example:
ViewRootImpl A --> decorView A --> ExpandMenuView A
ViewRootImpl B --> decorView A --> ExpandMenuView B
When rotating quickly, You want detach ExpandMenuView A from window. But actually,
You will detach ExpandMenuView B which should be a focus view.

Change-Id: If1e2e2799ca4373b0864a56571590f909060183a
Signed-off-by: jli119X <jianpingx.li@intel.com>
Reviewed-on: http://android.intel.com:8080/116543
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>