[PORT FROM MAIN]widi: Flush previous p2p scan results when enabling WFD feature
BZ: 118015

When WFD feature is OFF, p2p scan is done without including WFD IE in probe
requests. Some adapters like PTV3000 doesn't include WFD IE in their probe
responses if this field was not included in the request. Thus, they was not
considered as display devices.

When enabling the WFD feature, the devices list was not refreshed and adapters
responding without WFD IE were not proposed in the list to the user.

This patch flushes all previously discovered devices when enabling the WFD
feature and trigs a new scan in order to make a new discovery using WFD IE in
probe requests.

Change-Id: Ifd17866ef799842a763c9bf073554562774a8f28
Orig-Change-Id: Ia547aac66c2ac9d66a4731933bc9b3a7cf8e52b3
Category: device enablement
Domain: CWS.WIFI-Common
Origin: upstream
Upstream-Candidate: yes
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122954
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>