WiFi Display: Flush previous p2p scan results when enabling WFD feature
BZ: 118015

When WFD feature is OFF, p2p scan is done without including WFD IE in probe
requests. Some adapters like PTV3000 doesn't include WFD IE in their probe
responses if this field was not included in the request. Thus, they was not
considered as display devices.

When enabling the WFD feature, the devices list was not refreshed and adapters
responding without WFD IE were not proposed in the list to the user.

This patch flushes all previously discovered devices when enabling the WFD
feature and trigs a new scan in order to make a new discovery using WFD IE in
probe requests.

Change-Id: Ie3eb282419c2b47ed64d44e2dee7c2d70acac87a
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/116834
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>