[PORT FROM MAIN][AUDIO] Forward properly setForceUse requests from AudioService to Policy
BZ: 100472

There is a race condition when switching from speaker to BT_SCO between
the corresponding setForceUse calls. This results in random behaviors such
as ending on earpiece in the worst case scenario (BT_SCO->off information is
the last processed event which corresponds to setForceUse NONE)

This patch introduces a new method, reconsiderUsageForComm(), which will
reconsider the forced config switch in order to send the proper request
according to SCO/Speaker state.

Orig-Change-Id: Ibafde040ff4dbcfc263d8e4520eef43ba908c5fa

Change-Id: I46bc39d6ae6df5c8230591c34fb5314784d8767d
Signed-off-by: Piotr Diop <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/122731
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>