Wifi: Skip poor connection detection when disabled
BZ: 120429

In the WifiStateMachine, the state “VerifyingLinkState” was always processed
whatever the user preference set in advanced wifi settings. This state was
escaped only after receiving a good link detected message. Even if this step is
very short, the corresponding status was reported in the UI leading to a bad
user experience.

This patch skips this step and goes immediately to the next state if user has
disabled the poor link detection.

Change-Id: I4f5490dab5a7bca0ac06ebc3d1f2ec3845db494c
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/119209
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>