[PORT FROM MAIN] Wifi: Skip poor connection detection when disabled
BZ: 120429

In the WifiStateMachine, the state “VerifyingLinkState” was always processed
whatever the user preference set in advanced wifi settings. This state was
escaped only after receiving a good link detected message. Even if this step is
very short, the corresponding status was reported in the UI leading to a bad
user experience.

This patch skips this step and goes immediately to the next state if user has
disabled the poor link detection.

Category: AOSP improvement
Domain: WIFI-Core
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: I8243e9a228d6ec13e809ccbded1db92a05b4e136
Orig-Change-Id: I4f5490dab5a7bca0ac06ebc3d1f2ec3845db494c
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Signed-off-by: Loic Poulain <loicx.poulain@intel.com>
Reviewed-on: http://android.intel.com:8080/122474
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>