[PORT FROM MAIN JB-MR2] ConnectivityService: Adding IP default routes in Secondary Routing Tables
BZ: 124164

With ip route add, it's NOT possible to add a default Route
in a secondary routing table if a route to the Gateway is not
already installed in the Main Routing table.

This constraint is not taken into account in ConnectivityService
and this leads to routing issues.

Fix the issue by adding IP route to the gateway in the default
routing table.

Change-Id: I860d56cf1ac44529f31493fe6f989f64703e5a4d
Orig-Change-Id: Iff5c1af07a4264899b1a9066d16989e08025e582
Category: aosp improvement
Domain: Telephony-Base
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Zied Ben Hamouda <mohamed.zied.ben.hamouda@intel.com>