[PORT FROM MAIN] ConnectivityService: Adding IP default routes in Secondary Routing Tables
BZ: 124164

With ip route add, it's NOT possible to add a default Route
in a secondary routing table if a route to the Gateway is not
already installed in the Main Routing table.
Actually this constraint is not taken into account by ConnectivityService
and this can lead to routing issues.
Goal of this patch is to fix this issue.

Conflicts:
	services/java/com/android/server/ConnectivityService.java

Category: aosp improvement
Domain: AOSP-Framework-OS
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: Icc13ea8590a44391e1b37df18d503750a9a9caae
Orig-Change-Id: Iff5c1af07a4264899b1a9066d16989e08025e582
Signed-off-by: Zied Ben Hamouda <mohamed.zied.ben.hamouda@intel.com>
Reviewed-on: http://android.intel.com:8080/122171
Reviewed-by: Qin, Xiaokang <xiaokang.qin@intel.com>
Tested-by: Qin, Xiaokang <xiaokang.qin@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>