Shutdown: In airplane mode, shutdown modem using oem hooks
BZ: 123894

[DO NOT USPTREAM]

In airplane mode, radio is off but the modem can
still be powered on due to CWS clients accessing
the modem. If the platform shutdown is initiated
in this case, modem is not powered off properly.

On platform shutdown in airplane mode(radio off),
use oem hooks to power off the modem.

Change-Id: Id177c39f2ba881176f9ee600f849514a9f6cbcc6
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/119269
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>