Battery Service: Show lightning on battery icon only on  charging
BZ: 112501

Battery icon was showing lightning bolt icon when battery state
was 100% full and it was plain when the battery state was 99%
full. There should not be any variation in icon between the two
states since the battery status is the same, ie, 'full'. The aim
of this patch is to show the lightning bolt icon only when the
battery status is 'charging'.

Change-Id: I8f4ab51cbcd272e64232c0387aab4b68d2ee48f1
Signed-off-by: Saranya Gopal <saranya.gopal@intel.com>
Reviewed-on: http://android.intel.com:8080/120224
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>