widi: Changes needed in servicemanager to register widi service.
BZ: 125777

Widi needs to be registered as a service with servicemanager so that
it can be discovered by other processes at the native layer to talk to
widi service. This patch adds this capability, and uses INTEL_WIDI macro
to conditionally compile it in if the flag is passed during compilation.

This widi service is used by sigma agent and will be included only in userdebug
and eng build variants.

Change-Id: I08961aae39db6d5efccaab1757acedd92b8f2cfc
Signed-off-by: mamatha balguri <mamatha.balguri@intel.com>
Reviewed-on: http://android.intel.com:8080/120405
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Chowdhury, Rajneesh <rajneesh.chowdhury@intel.com>
Reviewed-by: Crabtree, Robert <robert.crabtree@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>