[PORT FROM MAIN] NetworkController: display the carrier name even when there is no data connection
BZ: 81455

The carrier name is replaced by 'NO INTERNET CONNECTION' when no
data connection available. It was requested by CCB to change this
behavior to have the carrier name displayed in all cases. This is
requested for some operator.

This patch changes the default behavior of the Android/UI to have
the carrier name displayed even when there is no data connection.
If the carrier name is empty the 'NO INTERNET CONNECTION' message
is displayed.

Domain: AOSP-Framework-SystemUI
Category: aosp improvement
Severity: medium
Origin: Internal
Upstream-Candidate: Yes
Orig-Change-Id: I168117ad1b04e842af787f72828ac01ba465554c
Change-Id: I55dfb66b82e67a9c4626423876aefb2c83ecca06
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/96313
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Tested-by: Shen, JianchunX <jianchunx.shen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>