[PORT FROM MAIN] Policy: Handling of Network Personalization PUK
BZ: 18871

Network personalisation is the feature were a SIM is restricted
to be used in a specific operator networks only. If an unsupported
SIM is used in the device, the user is asked to enter an unlock
code, so that the SIM can be used in the network. If the user enters
wrong PIN for a preprogrammed number of times, the device is PUK
locked and needs to be taken to the customer care. Android by default
supports entering the PIN, but doesn't inform the user if the device
is PUK locked.

This patch adds the handling of NETWORK PUK in the keyguard.

Change-Id: I7bef458ace57fbd29f0680ac71dcf39b5fc89b9e
Orig-Change-Id: I4e91529ddca2ee04d40c4cb1931ddd19b9770624
Orig-Change-Id: Ided3099fde34af2e63a3f3ddc590450817c647a2
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/121642
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Yao, Yong <yong.yao@intel.com>