WiFi Hotspot: Enable 5GHz and offer advanced settings to the user
BZ: 120839 123820

This patch allows the user to configure the WiFi Hotspot in an advanced way. The
user can now choose the band (2.4 or 5 GHz), change the hw mode ('bg', 'a' or
'ac'), set the 80211n mode and select the radio channel according to local
regulatory.
The UI is also ready to act on routing parameters like IP address and subnet
mask.

Technically this patch:
- Set the hostapd supplicant with hw, 80211n, channel and country parameters.
- Save and retrieve user choices with ConfigStore.
- Get from driver the allowed radio channels according to local regulatory.
- Reset the radio parameters when they are no more applicable in the new locale.

Change-Id: Ia1a870304b569b9006f3e5b88a0a1aa9e46cfe17
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/121671
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>