Wifi Display: add reconnectWifiDisplay in DisplayManager interface
BZ: 124677

Currently the reconnection to a wifidisplay is done by the WFD
settings activity. This force the user to stay in the setting
application until the reconnection is done

This patches adds a new method "reconnectWifiDisplay" in DisplayManager
interface that will handle the reconnection (ie. disconnection +
wait_scan + connection) in WifiDisplayController.
Thus as soon as the activity calls this method, reconnection will happen
even if the user leaves the WFD setting application.

Change-Id: I069a244c2e6de6753f97c646d49b5eb4cfc43b75
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>
Reviewed-on: http://android.intel.com:8080/121845
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Mauger, MatthieuX <matthieux.mauger@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>