[PORT FROM R42B-STABLE] Wifi Display: add reconnectWifiDisplay in DisplayManager interface
BZ: 124677

Currently the reconnection to a wifidisplay is done by the WFD
settings activity. This force the user to stay in the setting
application until the reconnection is done

This patches adds a new method "reconnectWifiDisplay" in DisplayManager
interface that will handle the reconnection (ie. disconnection +
wait_scan + connection) in WifiDisplayController.
Thus as soon as the activity calls this method, reconnection will happen
even if the user leaves the WFD setting application.

Category: aosp improvement
Domain: CWS.WIDI-Framework
Origin: internal
Upstream-Candidate: no

Orig-Change-Id: I069a244c2e6de6753f97c646d49b5eb4cfc43b75
Change-Id: Iec56833baf953dea18e6fa0f191f7c0f58db4839
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>
Reviewed-on: http://android.intel.com:8080/126114
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Mauger, MatthieuX <matthieux.mauger@intel.com>
Reviewed-by: Samoun, JacquesX <jacquesx.samoun@intel.com>
Reviewed-by: Garcia, Jeremie <jeremie.garcia@intel.com>
Reviewed-by: Tricot, HonoreX <honorex.tricot@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>