[PORT FROM MAIN]HTML5 Streaming: Hide NavigationBar when controlBar disappered.
BZ: 115879

When playback html5 video streaming in full screen mode,we can see NavigationBar
all the time which causes more power consumer. Need to hide NavigationBar
for saving power.
This patch also includes the fix that not show media controller when surface changed
if it is not in showing status.

Signed-off-by: Feng Xiao <fengx.xiao@intel.com>
Signed-off-by: ywan171 <yi.a.wang@intel.com>

Orig-Change-Id: I6b1d8656b3a3d56feb3285cd523999559201c2ff
Change-Id: Ib5cb000d3d6df3455dd5d8f4bba344c5f8133a59
Reviewed-on: http://android.intel.com:8080/121894
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>