[PORT FROM MAIN]HTML5: When enter to full screen mode, get video height and width
BZ: 103140

Before View::onMeasure is called, we need video height and width to
establish the real video display size or else it will be set to
default video surface size which equal to the screen size. So the
video frame seems to be stretched caused by the wrong video height.

Signed-off-by: Xiao,Feng <fengx.xiao@intel.com>
Signed-off-by: ywan171 <yi.a.wang@intel.com>

Orig-Change-Id: I0740bf58d682ebbc504cd2ed4a8c7e96ca6ba8a1
Change-Id: I2be863b874dd27bf9c55b4e9522f04522afedd1c
Reviewed-on: http://android.intel.com:8080/121927
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>