[PORT FROM MAIN]widi: Adapter reconnection upon AP just connected on another frequency
BZ: 113453 119803

Since Broadcom chip has some limitation to offer high video quality while used
in muli-role and multi-channel, the best solution is to use the same channel as
the access point when running in multi-role.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can image that
user first connects to video adapter and then connects to an access point. In
such situation, there is a  high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Fix Klocwork issues.

Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no,proprietary

Change-Id: I7bb22ed880797631afac842dfbb515877147a051
Orig-Change-Id: I0c7257fb72090cef80d643a0fe556790b0749252
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122826
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>