[PORT FROM MAIN] WiFi Display: Adapter reconnection upon AP just connected on another frequency
BZ: 113453 119803

Since Broadcom chip has some limitation to offer high video quality while used
in muli-role and multi-channel, the best solution is to use the same channel as
the access point when running in multi-role.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can image that
user first connects to video adapter and then connects to an access point. In
such situation, there is a  high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Fix Klocwork issues.

Category: AOSP improvement
Domain: WIFI-Core
Severity: high
Origin: internal
Upstream-Candidate: no

Change-Id: I0c7257fb72090cef80d643a0fe556790b0749252
Orig-Change-Id: Id802e75938615db841387a44c129c520502a73bf
Orig-Change-Id: Ib0144e03673fafbafb9a3a11b01fb7f38b5ad2ef
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: http://android.intel.com:8080/122826
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>