Tethering: fix the issue that cannot switch to mtp after enable/disable rndis.
BZ: 106269

If continuously enable/disable rndis, sometimes cannot switch to mtp after
disable rndis. That's because when user enables tether again, rndis disabled
process doesn't finish, then because rndis is not enabled and in process,
Tethering cannot do tether and update UI. Finally, function will switch to
rndis, but there's no chance to do tether and update UI again.
The solution is that when enable tether, if rndis is not enabled, don't set
mUsbTetherRequested to false. If set to false, when rndis is enabled, Tethering
cannot do tether and update UI again.

Change-Id: I7b7cde6c8e2a1474f014f2dc4818b3830fa2e02f
Signed-off-by: xiaochao <xiaochao.huang@intel.com>
Reviewed-on: http://android.intel.com:8080/123449
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Huang, Bo B <bo.b.huang@intel.com>
Reviewed-by: Wu, Hao <hao.wu@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Wang, Yu Y <yu.y.wang@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Wang, Ran <ran.wang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>