[PORT FROM MAIN] BatteryService: Shut down when battery is discharging on 0% with a cable connected
BZ: 47466

BatteryService: Shut down when battery is discharging on 0% with a cable connected.
If the platform consumes more power than the charger could supply and if the capacity
is zero, then shutdown the platform gracefully.

A seperate thread is spawned from user space at 0% which handles the shutdown event
at low battery and charger connected condition.
(BZ: 63823)

Orig-Change-Id: Ic102527ca7505a4209078f48fc9335c75ce60d60
Orig-Change-Id: Id69555826b5d9f713b01e749238712cefbd77de0

Change-Id: Ie5b09ba4b06933df363bb13873aeca1576de6310
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Reviewed-on: http://android.intel.com:8080/123700
Reviewed-by: Gopal, Saranya <saranya.gopal@intel.com>
Reviewed-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>